Custom C Programming Language Print F function clone

+++ _format_d.c
@@ -10,7 +10,7 @@
 
 int _format_d(va_list args)
 {
-	int n = va_arg(args, int);
+	long int n = va_arg(args, long int);
 	int res = 0;
 
 	if (n == 0)
@@ -23,8 +23,7 @@ int _format_d(va_list args)
 	{
 		_putchar('-');
 		res++;
-		/* Convert INT_MIN to positive without sign */
-		n = -(n + 1);
+		n = -(long int)(n + 1);
 	}
 
 	if (n < 0)
@@ -35,7 +34,7 @@ int _format_d(va_list args)
 	}
 
 	/* Output each digit of the number */
-	print_number(n, &res);
+	print_number((int)n, &res);
 
 	return (res);
 }
